Refactored frodo_red_subsky.c
Checked consistency of config/ERROR_CODES file.

Made sure that when comparing doubles, if precision is an issue, the gsl_fcmp() routine is used.

Made a header file for each routine and moved some variables from src/frodo_functions.h to there.

Replaced atoi() and atof() with strtol() and strtod().