Minor update:
Changed CRPIX reference values in frodo_red_rebin.c from 0.5 to 1 for fibre axes.

Changed comment of L2STATXC key from (CONTINUUM) to (CONT), as it was being truncated in the header in L2_exec.

Changed the structure of frodo_red_reformat.c so that the error_codes file and additional_keys_file are put into the header simultaneously.